+++ ui/src/app/core/accessories/types/windowcovering/windowcovering.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { ServiceTypeX } from '../../accessories.interfaces';
-import { WindowcoveringManageComponent } from './windowcovering.manage.component';
+import { WindowcoveringManageComponent } from '@/app/core/accessories/types/windowcovering/windowcovering.manage.component';
 
 @Component({
   selector: 'app-windowcovering',