+++ ui/src/app/modules/config-editor/config-editor.module.ts
@@ -4,10 +4,10 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { TranslateModule } from '@ngx-translate/core';
 import { MonacoEditorModule } from 'ngx-monaco-editor';
-import { ConfigEditorRoutingModule } from './config-editor-routing.module';
-import { ConfigEditorComponent } from './config-editor.component';
-import { ConfigEditorResolver } from './config-editor.resolver';
-import { ConfigRestoreBackupComponent } from './config-restore-backup/config.restore-backup.component';
+import { ConfigEditorRoutingModule } from '@/app/modules/config-editor/config-editor-routing.module';
+import { ConfigEditorComponent } from '@/app/modules/config-editor/config-editor.component';
+import { ConfigEditorResolver } from '@/app/modules/config-editor/config-editor.resolver';
+import { ConfigRestoreBackupComponent } from '@/app/modules/config-editor/config-restore-backup/config.restore-backup.component';
 
 @NgModule({
   declarations: [